requirements.txt
fastapi
uvicorn
langchain==0.2.16
langchain-community==0.2.16
langchain-google-genai==1.0.8
faiss-cpu
sentence-transformers
pymupdf
streamlit
python-dotenv
requests
sqlalchemy
psycopg2-binary
supabase